Move and adjust the test to QualifierCircularDependency etc
